# Conflicts:
#	RevevueApp/pom.xml
#	RevevueApp/src/main/java/fr/pentagon/ugeoverflow/controllers/rest/QuestionController.java
#	RevevueApp/src/main/java/fr/pentagon/ugeoverflow/service/QuestionService.java
#	RevevueApp/src/main/java/fr/pentagon/ugeoverflow/service/ReviewService.java

+++ RevevueApp/src/main/java/fr/pentagon/ugeoverflow/controllers/dtos/responses/ReviewQuestionResponseDTO.java
@@ -5,5 +5,5 @@ import jakarta.annotation.Nullable;
 import java.util.Date;
 import java.util.List;
 
-public record ReviewQuestionResponseDTO(long id, String author, Date creationDate, String content, @Nullable String citedCode, int upvotes, int downvotes, List<ReviewResponseDTO> reviews) {
+public record ReviewQuestionResponseDTO(long id, String author, Date creationDate, String content, @Nullable String citedCode, long upvotes, long downvotes, List<ReviewResponseDTO> reviews) {
 }
